social_auth_service/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
bcrypt==4.2.1
requests==2.32.3
python-dotenv==1.0.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1
PyJWT==2.10.1
